requirements.txt
moneylib
django